[oap-cache/oap]Add oap-perf-suite for oap-cache/oap (#1217)
* add oap-perf-suite for OAP branch-0.8

* Clean oap file format codes
