+++ oap-cache/oap/pom.xml
@@ -625,7 +625,7 @@
         <version>3.2.0</version>
         <configuration>
           <descriptors>
-            <descriptor>../../src/test/assembly/test-jar-with-dependencies.xml</descriptor>
+            <descriptor>src/test/assembly/test-jar-with-dependencies.xml</descriptor>
           </descriptors>
         </configuration>
         <executions>
