+++ oap-cache/oap/src/test/assembly/test-jar-with-dependencies.xml
